Creating limitations to saving to a directory that doesnt exist, creating other threads for Auto sorting and Sorting and other minor stuff

+++ main.py
@@ -4,6 +4,8 @@ from tkinter.scrolledtext import ScrolledText
 import shutil
 import sys
 import os
+import time
+from threading import Thread
 from configparser import ConfigParser
 
 # zaczytywanie pliku konfikuracyjnego
@@ -40,7 +42,7 @@ class Main:
     def __init__(self):   
     
         self.mainWin = tk.Tk(className=title)
-        self.mainWin.geometry('800x500')
+        self.mainWin.geometry('750x400')
         
         self.upperMenu = tk.Menu(self.mainWin)
         self.menuSettings = tk.Menu(self.upperMenu, tearoff=0)
@@ -53,6 +55,11 @@ class Main:
         self.upperMenu.configure(bg='#ffffff')
         
         self.Title = tk.Label(self.mainWin, text='Simply Sort', font=('Caladea', 24)).pack()
+        
+        self.PathsReadOnly = tk.Entry(self.mainWin, state='normal', width=35)
+        self.PathsReadOnly.pack(pady=5)
+        self.PathsReadOnly.insert(tk.END, savedSource)
+        self.PathsReadOnly.configure(state='readonly')
         
         # główny kontener
         self.container = tk.Frame(self.mainWin)
@@ -73,9 +80,29 @@ class Main:
         self.description = tk.Label(self.containerDesc, text='Schemas:', justify='left', font=('David Libre', 16))
         self.description.grid(row=0, column=0, sticky=tk.E+tk.W)
         
-        # schemty
-        self.description = tk.Label(self.containerDesc, text='Lorem ipsum dolor sit amet,\nconsectetur adipiscing elit.\nCurabitur aliquet turpis a\ntincidunt consequat. Integer\nfinibus mauris in commodo\npellentesque. Pellentesque\nauctor neque id arcu hendrerit,\nid iaculis felis faucibus.\nNam vulputate dolor neque,\nnon iaculis tortor tempus vitae.', justify='left', font=('David Libre', 11))
-        self.description.grid(row=1, column=0, sticky=tk.E+tk.W)
+        # schema 1
+        self.PathsReadOnly = tk.Entry(self.containerDesc, state='normal', width=35)
+        self.PathsReadOnly.grid(row=1, column=0, pady=5)
+        self.PathsReadOnly.insert(tk.END, savedSchema)
+        self.PathsReadOnly.configure(state='readonly')
+        
+        # Path2
+        self.PathsReadOnly1 = tk.Entry(self.containerDesc, state='normal', width=35)
+        self.PathsReadOnly1.grid(row=2, column=0, pady=5)
+        self.PathsReadOnly1.insert(tk.END, savedSchema1)
+        self.PathsReadOnly1.configure(state='readonly')
+        
+        # Path3
+        self.PathsReadOnly2 = tk.Entry(self.containerDesc, state='normal', width=35)
+        self.PathsReadOnly2.grid(row=3, column=0, pady=5)
+        self.PathsReadOnly2.insert(tk.END, savedSchema2)
+        self.PathsReadOnly2.configure(state='readonly')
+        
+        # Path4
+        self.PathsReadOnly3 = tk.Entry(self.containerDesc, state='normal', width=35)
+        self.PathsReadOnly3.grid(row=4, column=0, pady=5)
+        self.PathsReadOnly3.insert(tk.END, savedSchema3)
+        self.PathsReadOnly3.configure(state='readonly')
         
         # kontener na przyciski
         self.containerButtons = tk.Frame(self.containerFunc)
@@ -83,8 +110,8 @@ class Main:
         self.containerFunc.columnconfigure(1, weight=1)
         
         # przycisk "SORT!"
-        self.sortButton = tk.Button(self.containerButtons, width=7, height=1, text='SORT!', font=('David Libre', 15), command=self.sortItems, disabledforeground="#363636")
-        self.sortButton.grid(row=0, column=0, padx=5)
+        self.sortButton = tk.Button(self.containerButtons, width=7, height=1, text='SORT!', font=('David Libre', 15), command=self.sortStart, disabledforeground="#363636")
+        self.sortButton.grid(row=0, column=0, padx=2)
         
         # ramka z przyciskiem do autosortowania
         self.autoSGrid = tk.Frame(self.containerButtons)
@@ -108,7 +135,7 @@ class Main:
         self.containerPaths.columnconfigure(1, weight=1)
         
         # tytuł "Your Paths"
-        self.PathsTitle = tk.Label(self.containerPaths, text="Your Paths:", font=('David Libre', 11))
+        self.PathsTitle = tk.Label(self.containerPaths, text="Your Paths:", font=('David Libre', 16))
         self.PathsTitle.grid(row=0, column=0)
         
         # Path 1
@@ -150,38 +177,67 @@ class Main:
         sys.stderr = logger
         
         self.mainWin.mainloop()
+    
+    def sortStart(self):
+        self.sortThread = Thread(target = self.sortItems)
+        self.sortThread.start()
                 
     def sortItems(self):
-        print("Starting sorting...\n")
-        
-        for file in os.listdir(savedSource):
-            if file.endswith(savedSchema1):
-                shutil.move(f'{savedSource}/{file}', savedtarget1)
-                print(f'Sorting to {savedtarget1} directory finished succesfuly')
-            elif file.endswith(savedSchema2):
-                shutil.move(f'{savedSource}/{file}', savedtarget2)
-                print(f'Sorting to {savedtarget2} directory finished succesfuly')
-            elif file.endswith(savedSchema3):
-                shutil.move(f'{savedSource}/{file}', savedtarget3)
-                print(f'Sorting to {savedtarget3} directory finished succesfuly')
+        if os.path.isdir(savedSource) is False:
+            print('Path that you entered cannot be found! :(')
+        else:
+            dir = os.listdir(savedSource)
+            if len(dir) == 0:
+                print('Your sorting folder is empty!')
+            else:
+                print("--------------------------------------------------------------------------------\nStarting sorting...\n--------------------------------------------------------------------------------")
+                for file in os.listdir(savedSource):
+                    if file.endswith(savedSchema):
+                        if os.path.exists(savedtarget) is True:
+                            shutil.move(f'{savedSource}/{file}', savedtarget)
+                            print(f'Sorting to {savedtarget.upper()} directory finished succesfuly')
+                        elif os.path.exists(savedtarget) is False:
+                            print(f'Path {savedtarget.upper()} does not exist!')
+                    elif file.endswith(savedSchema1):
+                        if os.path.exists(savedtarget1) is True:
+                            shutil.move(f'{savedSource}/{file}', savedtarget1)
+                            print(f'Sorting to {savedtarget1.upper()} directory finished succesfuly')
+                        elif os.path.exists(savedtarget1) is False:
+                            print(f'Path {savedtarget1.upper()} does not exist!')
+                    elif file.endswith(savedSchema2):
+                        if os.path.exists(savedtarget2) is True:
+                            shutil.move(f'{savedSource}/{file}', savedtarget2)
+                            print(f'Sorting to {savedtarget2.upper()} directory finished succesfuly')
+                        elif os.path.exists(savedtarget2) is False:
+                            print(f'Path {savedtarget2.upper()} does not exist!')
+                    elif file.endswith(savedSchema3):
+                        if os.path.exists(savedtarget3) is True:
+                            shutil.move(f'{savedSource}/{file}', savedtarget3)
+                            print(f'Sorting to {savedtarget3.upper()} directory finished succesfuly')
+                        elif os.path.exists(savedtarget3) is False:
+                            print(f'Path {savedtarget3.upper()} does not exist!')
             
-        print('\nSorting finished succefully!')
+                print('--------------------------------------------------------------------------------\nSorting finished succefully!\n--------------------------------------------------------------------------------')
         
     def autoSort(self):
+        
+        self.thread = Thread(target=self.checkingForFiles)
+        
         self.sortingVar += 1
         if self.sortingVar % 2 != 0:
             self.autoSorting['text'] = "ON "
             self.autoSorting.configure(bg='#00ff00')
             self.sortButton['state'] = 'disabled'
             self.sortButton['text'] = ' ... '
-            print('Autosorting ON')
+            print('--------------------------------------------------------------------------------\nAutosorting ON\n--------------------------------------------------------------------------------')
+            self.thread.start()
         elif self.sortingVar % 2 == 0:
             self.autoSorting['text'] = "OFF"
             self.autoSorting.configure(bg='#ff0000')
             self.sortButton['state'] = 'normal'
             self.sortButton['text'] = 'SORT!'
-            print('Autosorting OFF')
-    
+            self.temp += 2000000
+
     def setPaths(self):
         self.mainWin.destroy()
         os.system('python.exe paths.py')
@@ -189,5 +245,23 @@ class Main:
     def setOther(self):
         self.mainWin.destroy()
         os.system('python.exe other.py')
+        
+    def checkingForFiles(self):
+        self.temp = 0
+        while self.temp<1000000:
+            if len(os.listdir(savedSource)) == 0:
+                time.sleep(2)
+                print('Looking for some files...')
+                self.temp += 1
+            elif len(os.listdir(savedSource)) > 0:
+                self.sortItems()
+                time.sleep(2)
+                self.temp += 1
+        else:
+            self.autoSorting['text'] = "OFF"
+            self.autoSorting.configure(bg='#ff0000')
+            self.sortButton['state'] = 'normal'
+            self.sortButton['text'] = 'SORT!'
+            print('--------------------------------------------------------------------------------\nAutosorting OFF\n--------------------------------------------------------------------------------')
         
 Main()
